backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.18.1
python-multipart==0.0.9
python-dotenv==1.0.0
fastapi-mail==1.4.1